+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
@@ -17,44 +18,46 @@ class RoleController extends Controller
      */
     public function index(Request $request)
     {
+        if(Auth::user()->can('role.view')) 
+        {
+            $roles = DB::table('roles')->orderBy('id', 'DESC')->get();
 
-        $roles = DB::table('roles')->orderBy('id', 'DESC')->get();
-
-        if($request->ajax()) {
-            return DataTables::of($roles)
-            ->addColumn('created_at', function($role) {
-                return date("M d, Y H:i:s", strtotime($role->created_at));
-            })
-            ->addColumn('action', function($role) {
-                return '<div class="dropdown">
-                            <button class="btn btn-primary btn-xs dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
-                            Action
-                            </button>
-                            <ul class="dropdown-menu">
-                                <li>
-                                    <a href="' . route('roles.edit', $role->id) . '" class="dropdown-item">
-                                        <i class="fas fa-edit"></i>
-                                            Edit
-                                    </a>
-                                </li>
-                                <li>
-                                    <form action="' . route('roles.destroy', $role->id) . '" method="post" class="ajax-delete">'
-                                        . csrf_field() 
-                                        . method_field('DELETE') 
-                                        . '<button type="button" class="btn-remove dropdown-item">
-                                                <i class="fas fa-trash-alt"></i>
-                                                    Delete
-                                            </button>
-                                    </form>
-                                </li>
-                            </ul>
-                    </div>';
-            })
-            ->rawColumns(['created_at', 'action'])
-            ->make(true);
+            if($request->ajax()) {
+                return DataTables::of($roles)
+                ->addColumn('created_at', function($role) {
+                    return date("M d, Y H:i:s", strtotime($role->created_at));
+                })
+                ->addColumn('action', function($role) {
+                    return '<div class="dropdown">
+                                <button class="btn btn-primary btn-xs dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
+                                Action
+                                </button>
+                                <ul class="dropdown-menu">
+                                    <li>
+                                        <a href="' . route('roles.edit', $role->id) . '" class="dropdown-item">
+                                            <i class="fas fa-edit"></i>
+                                                Edit
+                                        </a>
+                                    </li>
+                                    <li>
+                                        <form action="' . route('roles.destroy', $role->id) . '" method="post" class="ajax-delete">'
+                                            . csrf_field() 
+                                            . method_field('DELETE') 
+                                            . '<button type="button" class="btn-remove dropdown-item">
+                                                    <i class="fas fa-trash-alt"></i>
+                                                        Delete
+                                                </button>
+                                        </form>
+                                    </li>
+                                </ul>
+                        </div>';
+                })
+                ->rawColumns(['created_at', 'action'])
+                ->make(true);
+            }
+            return view('roles.index');
         }
-        return view('roles.index');
-
+        abort(403);
     }
 
     /**
@@ -64,9 +67,13 @@ class RoleController extends Controller
      */
     public function create()
     {
-        $all_permissions = DB::table('permissions')->get();
-        $permission_groups = DB::table('permissions')->select('group_name')->distinct()->get();
-        return view("roles.create", compact('permission_groups', 'all_permissions'));
+        if(Auth::user()->can('role.create')) 
+        {
+            $all_permissions = DB::table('permissions')->get();
+            $permission_groups = DB::table('permissions')->select('group_name')->distinct()->get();
+            return view("roles.create", compact('permission_groups', 'all_permissions'));
+        }
+        abort(403);
     }
 
     /**
@@ -109,10 +116,14 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
-        $permission_groups = DB::table('permissions')->select('group_name')->distinct()->get();
-        $all_permissions = DB::table('permissions')->get();
-        return view("roles.edit", compact('role', 'permission_groups', 'all_permissions'));
+        if(Auth::user()->can('role.edit')) 
+        {
+            $role = Role::findById($id);
+            $permission_groups = DB::table('permissions')->select('group_name')->distinct()->get();
+            $all_permissions = DB::table('permissions')->get();
+            return view("roles.edit", compact('role', 'permission_groups', 'all_permissions'));
+        }
+        abort(403);
     }
 
     /**
@@ -163,15 +174,19 @@ class RoleController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        $role = Role::findById($id);
-        if(!is_null($role)){
-            $role->delete();
+        if(Auth::user()->can('role.delete')) 
+        {
+            $role = Role::findById($id);
+            if(!is_null($role)){
+                $role->delete();
+            }
+    
+            if (!$request->ajax()) {
+                return back()->with('success', 'Information has been deleted!');
+            } else {
+                return response()->json(['result' => 'success', 'message' => 'Information has been deleted sucessfully']);
+            }
         }
-
-        if (!$request->ajax()) {
-            return back()->with('success', 'Information has been deleted!');
-        } else {
-            return response()->json(['result' => 'success', 'message' => 'Information has been deleted sucessfully']);
-        }
+        abort(403);
     }
 }
